feat: support System.iterator for-loops
Allow iterating over the content of the map using a for() loop like this:
```javascript
for(const [key, value] of myOrderedMap) {
   ...
}
```

Also introduce initial jest configuration with a basic test of the new feature.
NOTE: this change does not include adding missing tests for all other features of the OrderedMap, as it is beyond the scope of this update.

This updated is in relation to [this suggestion](https://github.com/marijnh/orderedmap/issues/16) requested on the original [repo](https://github.com/marijnh/orderedmap).